<refactor> LadderGameInfo dto 패키지로 변경
